takes pri range down a notch.

+++ CiCa.h
@@ -21,7 +21,7 @@ struct CiCa
         bool Set;
     };
 
-    CiCa(int64_t a_Priority)
+    CiCa(int32_t a_Priority)
     : m_W(&m_Ts[0])
     , m_R(&m_Ts[1])
     , m_Priority(a_Priority)
@@ -45,5 +45,5 @@ struct CiCa
     End m_Ts[2];
     End* m_W;
     End* m_R;
-    int64_t m_Priority;
+    int32_t m_Priority;
 };
